Refactoring of public server API
