+++ platform/dolphin-platform-server-javaee/src/main/resources/META-INF/services/com.canoo.platform.server.spi.components.ManagedBeanFactory
