+++ platform/dolphin-platform-server-spring/src/main/resources/META-INF/services/com.canoo.platform.server.spi.components.ManagedBeanFactory
