+++ platform/dolphin-platform-server/src/main/java/com/canoo/platform/server/spi/ServerCoreComponents.java
@@ -15,11 +15,8 @@
  */
 package com.canoo.platform.server.spi;
 
-import com.canoo.dp.impl.server.beans.ManagedBeanFactory;
 import com.canoo.platform.core.PlatformConfiguration;
 import org.apiguardian.api.API;
-
-import javax.servlet.ServletContext;
 
 import static org.apiguardian.api.API.Status.EXPERIMENTAL;
 
@@ -32,32 +29,11 @@ import static org.apiguardian.api.API.Status.EXPERIMENTAL;
 public interface ServerCoreComponents {
 
     /**
-     * Returns the servlet context of the underlying servlet engine.
-     *
-     * @return the servlet context
-     */
-    ServletContext getServletContext();
-
-    /**
      * Returns the configuration of the Dolphin Platform
      *
      * @return the configuration
      */
     PlatformConfiguration getConfiguration();
-
-    /**
-     * Returns the classpath scanner that is used by Dolphin Platform
-     *
-     * @return the classpath scanner
-     */
-    ClasspathScanner getClasspathScanner();
-
-    /**
-     * Returns the bean factory that is used by the Dolphin Platform to create managed beans based on the underlying platform
-     *
-     * @return the bean factory
-     */
-    ManagedBeanFactory getManagedBeanFactory();
 
     /**
      * The {@link ServerCoreComponents} can hold implementations for a given service and share this between several modules (see {@link ServerModule}). This method must be used to add such an instance of a service to the core components.
