+++ platform/dolphin-platform-server/src/main/java/com/canoo/platform/server/spi/components/ClasspathScanner.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.platform.server.spi;
+package com.canoo.platform.server.spi.components;
 
 import org.apiguardian.api.API;
